requirements.txt
python-binance
cryptography
ccxt
requests
discord.py
boto3
